+++ akeyless/retry_resource.go
@@ -10,29 +10,33 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-// resource-level retry default backoff values.
+// Resource-level retry defaults (longer intervals than provider).
 const (
-	resourceRetryDefaultMaxRetries         = 3
-	resourceRetryDefaultIntervalSeconds    = 10.0
-	resourceRetryDefaultMaxIntervalSeconds = 180.0
-	resourceRetryDefaultMultiplier         = 1.5
+	resourceRetryDefaultIntervalSeconds   = 10.0
+	resourceRetryDefaultMaxBackoffSeconds = 180.0
 )
 
-// resourceRetrySchema is the optional resource-level retry block. When set, the
-// resource's API calls retry with these settings instead of the provider defaults.
+// resourceRetrySchema returns the optional resource-level retry {} block schema.
+// When set, overrides provider HTTP retry for this resource.
 func resourceRetrySchema() *schema.Schema {
 	return &schema.Schema{
 		Type:        schema.TypeList,
 		Optional:    true,
 		MaxItems:    1,
-		Description: "Optional resource-level retry. When set, overrides provider HTTP retry for this resource.",
+		Description: "Optional resource-level retry. When set, overrides provider HTTP retry.",
 		Elem: &schema.Resource{
 			Schema: map[string]*schema.Schema{
-				"retry_on_messages": {
+				"max_retries": {
+					Type:        schema.TypeInt,
+					Optional:    true,
+					Default:     defaultMaxRetries,
+					Description: "Maximum retries after the first attempt.",
+				},
+				"retry_on_status_codes": {
 					Type:        schema.TypeList,
 					Optional:    true,
-					Description: "Extra error messages that trigger a retry (matched as regex or plain text).",
-					Elem:        &schema.Schema{Type: schema.TypeString},
+					Description: "HTTP status codes that trigger a retry. Defaults to 429, 500, 502, 503, 504.",
+					Elem:        &schema.Schema{Type: schema.TypeInt},
 				},
 				"interval_seconds": {
 					Type:        schema.TypeFloat,
@@ -40,32 +44,31 @@ func resourceRetrySchema() *schema.Schema {
 					Default:     resourceRetryDefaultIntervalSeconds,
 					Description: "Initial backoff interval in seconds.",
 				},
-				"max_interval_seconds": {
+				"max_backoff_seconds": {
 					Type:        schema.TypeFloat,
 					Optional:    true,
-					Default:     resourceRetryDefaultMaxIntervalSeconds,
+					Default:     resourceRetryDefaultMaxBackoffSeconds,
 					Description: "Maximum backoff interval in seconds.",
 				},
 				"multiplier": {
 					Type:        schema.TypeFloat,
 					Optional:    true,
-					Default:     resourceRetryDefaultMultiplier,
+					Default:     defaultMultiplier,
 					Description: "Exponential backoff multiplier.",
 				},
-				"max_retries": {
-					Type:        schema.TypeInt,
+				"retry_on_messages": {
+					Type:        schema.TypeList,
 					Optional:    true,
-					Default:     resourceRetryDefaultMaxRetries,
-					Description: "Maximum number of retries after the first attempt.",
+					Description: "Extra error messages that trigger a retry (matched as regex).",
+					Elem:        &schema.Schema{Type: schema.TypeString},
 				},
 			},
 		},
 	}
 }
 
-// EnableResourceRetry adds the retry block to the resource and makes each CRUD hook
-// use a resource-configured retry client when the block is set. With no retry block,
-// the resource keeps using the provider client unchanged.
+// EnableResourceRetry adds the retry {} block to a resource and wraps CRUD hooks.
+// When a resource has a retry block, its operations use a custom-configured client.
 func EnableResourceRetry(r *schema.Resource) {
 	if r == nil {
 		return
@@ -75,7 +78,6 @@ func EnableResourceRetry(r *schema.Resource) {
 	}
 	if _, exists := r.Schema["retry"]; !exists {
 		s := resourceRetrySchema()
-		// Create-only resources (no Update) require ForceNew on every schema attribute.
 		if r.Update == nil && r.UpdateContext == nil {
 			s.ForceNew = true
 		}
@@ -144,9 +146,8 @@ func EnableResourceRetry(r *schema.Resource) {
 	}
 }
 
-// resourceRetryDeps returns the provider meta to use for this operation. When the
-// resource has a retry {} block, it returns a copy whose client retries per the
-// block's settings; otherwise it returns m unchanged (provider retry applies).
+// resourceRetryDeps returns provider meta with a custom retry client when
+// a retry {} block is present; otherwise returns the original meta unchanged.
 func resourceRetryDeps(d *schema.ResourceData, m interface{}) (interface{}, error) {
 	cfg, ok, err := resourceRetryConfigFromData(d)
 	if err != nil {
@@ -154,39 +155,44 @@ func resourceRetryDeps(d *schema.ResourceData, m interface{}) (interface{}, erro
 	}
 	pm, isMeta := m.(*providerMeta)
 	if !ok || !isMeta || pm == nil {
-		return m, nil
+		return m, nil // no retry block set → use provider client
 	}
-
-	cp := *pm // shallow copy shares the token
+	// resource retry {} set → new client with resource cfg
+	cp := *pm
 	cp.client = buildAPIClient(pm.apiGwAddress, newRetryTransport(http.DefaultTransport, cfg))
 	return &cp, nil
 }
 
-// resourceRetryConfigFromData builds a transport retry config from the resource's
-// retry {} block. ok is false when the block is absent.
+// resourceRetryConfigFromData parses the resource's retry {} block.
+// Returns ok=false when block is absent.
 func resourceRetryConfigFromData(d *schema.ResourceData) (cfg retryConfig, ok bool, err error) {
 	raw := d.Get("retry").([]interface{})
 	if len(raw) == 0 || raw[0] == nil {
 		return retryConfig{}, false, nil
 	}
-	m := raw[0].(map[string]interface{})
 
-	// Start from provider defaults so the resource still retries busy status codes
-	// and rate-limit responses; then apply the block's overrides.
+	m := raw[0].(map[string]interface{})
 	cfg = defaultRetryConfig()
+
 	if v, ok := m["max_retries"].(int); ok {
 		cfg.MaxRetries = v
 	}
 	if v, ok := m["interval_seconds"].(float64); ok {
 		cfg.IntervalSeconds = v
 	}
-	if v, ok := m["max_interval_seconds"].(float64); ok {
+	if v, ok := m["max_backoff_seconds"].(float64); ok {
 		cfg.MaxBackoffSeconds = v
 	}
 	if v, ok := m["multiplier"].(float64); ok {
 		cfg.Multiplier = v
 	}
-
+	if codes, ok := m["retry_on_status_codes"].([]interface{}); ok && len(codes) > 0 {
+		parsed := make([]int, 0, len(codes))
+		for _, c := range codes {
+			parsed = append(parsed, c.(int))
+		}
+		cfg.RetryOnStatusCodes = statusCodeSet(parsed)
+	}
 	if patterns, ok := m["retry_on_messages"].([]interface{}); ok {
 		regexes := make([]*regexp.Regexp, 0, len(patterns))
 		for _, p := range patterns {
@@ -202,5 +208,6 @@ func resourceRetryConfigFromData(d *schema.ResourceData) (cfg retryConfig, ok bo
 		}
 		cfg.ErrorMessageRegex = regexes
 	}
+
 	return cfg, true, nil
 }
